+++ src/main/kotlin/xyz/nibblz/galapagos/data/BlueprintLootPreview.kt
@@ -1,10 +1,10 @@
 package xyz.nibblz.galapagos.data
 
 data class BlueprintLootPreview(
-    var newCosmeticChance: Float = 0.0f,
+    var newCosmeticChance: Double = 0.0,
     var currentTrophies: Int = 0,
     var totalTrophies: Int = 0,
-    var newRepChance: Float = 0.0f,
+    var newRepChance: Double = 0.0,
     var currentRep: Int = 0,
     var totalRep: Int = 0
 )
