+++ src/main/kotlin/xyz/nibblz/galapagos/data/Cosmetic.kt
@@ -1,7 +1,7 @@
 package xyz.nibblz.galapagos.data
 
 data class Cosmetic(
-    var chance: Float,
+    var chance: Double,
     var isOwned: Boolean,
     var trophies: Int,
     var rep: Int
